Altered dispatch to use an array (slice) of handlers rather than a map.
This makes handler execution order consistent with the order of handler
registration and allows for a default "match all unmatched commands"
handler which is only executed if no other handlers have matched.

This should not decrease the chatbot framework's efficiency since the
previous map implementation performed a linear loop through its handlers
regardless, making an array container more suited for the task.

The future performance implications that I could think of would be if a
"delete handler" operation was added although this could also perform a
linear search through what will hopefully be a limited number of
handlers.